init/requirements.txt
dnspython==2.3.0
pymongo==4.5.0
python-dotenv==0.21.1
